Fluffle.Notebooks/requirements.txt
tensorflow==2.6.0
pandas==1.0.5
numpy==1.19.5
scikit-learn==0.23.1
more-itertools==8.8.0
Pillow==8.3.2
fastapi==0.70.0
python-multipart==0.0.5
joblib==1.0.0
b2sdk==1.12.0
uvicorn==0.14.0
pyaml==20.4.0
